+++ src/exec/exec_utils.c
@@ -6,54 +6,11 @@
 /*   By: JFikents <Jfikents@student.42Heilbronn.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/15 17:15:16 by JFikents          #+#    #+#             */
-/*   Updated: 2024/07/08 18:25:32 by JFikents         ###   ########.fr       */
+/*   Updated: 2024/07/08 20:55:59 by JFikents         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-int	get_fd(t_token *redir)
-{
-	int		fd;
-
-	fd = -1;
-	if (redir->type == REDIR_TO || redir->type == APPEND_TO)
-		fd = 1;
-	else if (redir->type == REDIR_FROM || redir->type == HERE_DOC)
-		fd = 0;
-	if (ft_isdigit(redir->value[0]))
-		fd = ft_atoi(redir->value);
-	return (fd);
-}
-
-static int	set_other_redirections(t_token *redir)
-{
-	char	*file;
-	int		original_fd;
-	int		fd;
-
-	while (redir != NULL)
-	{
-		original_fd = get_fd(redir);
-		file = redir->next->value;
-		if (redir->type == REDIR_TO)
-			fd = open(file, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-		else if (redir->type == APPEND_TO)
-			fd = open(file, O_WRONLY | O_CREAT | O_APPEND, 0644);
-		else if (redir->type == REDIR_FROM)
-			fd = open(file, O_RDONLY, 0644);
-		else if (redir->type == HERE_DOC)
-			fd = open(HEREDOC_FILE, O_RDONLY, 0644);
-		if (fd == -1)
-			return (ft_putstr_fd("minishell: Error opening file\n", 2), 1);
-		if (dup2(fd, original_fd) == -1)
-			return (ft_printf_fd(2, "minishell: %d: Bad file descriptor\n",
-					original_fd), ft_close(&fd), 1);
-		ft_close(&fd);
-		redir = redir->next->next;
-	}
-	return (0);
-}
 
 void	free_cmd(t_cmd **cmd)
 {
@@ -73,29 +30,6 @@ void	free_cmd(t_cmd **cmd)
 		ft_close(&tmp->pipe[PIPE_FD_WRITE]);
 		ft_free_n_null((void **)&tmp);
 	}
-}
-
-int	do_all_redirections(t_cmd *cmd)
-{
-	int	ret;
-
-	if (check_if_heredoc(cmd->redirects))
-		return (EXIT_FAILURE);
-	if (cmd->pipe[PIPE_FD_READ] != 0)
-	{
-		ret = setup_in_pipe(cmd->pipe);
-		if (ret == -1)
-			return (EXIT_FAILURE);
-	}
-	if (cmd->pipe[PIPE_FD_WRITE] != 0)
-	{
-		ret = setup_out_pipe(cmd->pipe);
-		if (ret == -1)
-			return (EXIT_FAILURE);
-	}
-	if (set_other_redirections(cmd->redirects))
-		return (EXIT_FAILURE);
-	return (EXIT_SUCCESS);
 }
 
 static int	is_builtin(const char *cmd)
@@ -120,31 +54,24 @@ static int	is_builtin(const char *cmd)
 	return (false);
 }
 
-pid_t	execute_cmd(t_cmd *cmd)
+char	**transform_to_array(t_token *token)
 {
-	pid_t		pid;
-	int			exit_code;
-	int			builtin;
+	char	**argv;
+	int		argc;
 
-	builtin = is_builtin(cmd->argv[0]);
-	if (builtin == true)
+	argc = 1;
+	while (token->next != NULL && ++argc)
+		token = token->next;
+	argv = ft_calloc(argc + 1, sizeof(char *));
+	if (argv == NULL)
+		return (NULL);
+	while (token->prev != NULL)
+		token = token->prev;
+	argc = 0;
+	while (token != NULL)
 	{
-		exit_code = exec_builtins(cmd);
-		set_last_process_exit_code(exit_code);
-		return (BUILTIN_EXECUTED);
+		argv[argc++] = ft_strdup(token->value);
+		token = token->next;
 	}
-	if (builtin == -1)
-		return (ft_printf_fd(2, ERROR_MSG, cmd->argv[0], E_ALLOC),
-			EXIT_FAILURE);
-	pid = fork();
-	if (pid == -1)
-		return (ft_printf_fd(2, ERROR_MSG, cmd->argv[0], "Error creating fork"),
-			EXIT_FAILURE);
-	if (pid == 0)
-	{
-		if (do_all_redirections(cmd))
-			return (EXIT_FAILURE);
-		ft_execve(cmd);
-	}
-	return (pid);
+	return (argv);
 }
